+++ SDK/include/luabind/detail/implicit_cast.hpp
@@ -41,6 +41,6 @@ namespace luabind { namespace detail
 	// the this-pointer is const, there's no problem, since the non-const function
 	// will not match at all.
 
-	LUABIND_API int implicit_cast(const class_rep*, LUABIND_TYPE_INFO const&, ptrdiff_t&);
+	LUABIND_API int implicit_cast(const class_rep*, LUABIND_TYPE_INFO const&, int&);
 
 }}
